+++ sdk/src/com/appnexus/opensdk/mediatedviews/AdMobBanner.java
@@ -50,6 +50,26 @@ public class AdMobBanner implements MediatedBannerAdView {
     public void destroy() {
         if (adView != null) {
             adView.destroy();
+            adView=null;
         }
+    }
+
+    @Override
+    public void onPause() {
+        if(adView!=null) {
+            adView.onPause();
+        }
+    }
+
+    @Override
+    public void onResume() {
+        if(adView!=null) {
+            adView.onResume();
+        }
+    }
+
+    @Override
+    public void onDestroy() {
+        destroy();
     }
 }