+++ sdk/src/com/appnexus/opensdk/mediatedviews/AdMobInterstitial.java
@@ -59,6 +59,28 @@ public class AdMobInterstitial implements MediatedInterstitialAdView {
     public void destroy() {
         if (adView != null) {
             adView.destroy();
+            adView=null;
+        }
+    }
+
+    @Override
+    public void onPause() {
+        if(adView!=null) {
+            adView.onPause();
+        }
+    }
+
+    @Override
+    public void onResume() {
+        if(adView!=null) {
+            adView.onResume();
+        }
+    }
+
+    @Override
+    public void onDestroy() {
+        if(adView!=null) {
+            adView.onDestroy();
         }
     }
 }